EE-16843 Split the unit tests for the PassStatisticsAccumulator into two, one for the simple cases of calling the accumulate method once, the other for cases calling it multiple times
